Added 64-bit getrs (#255)
* Added 64-bit getrs APIs

* Testing 64-bit getrs APIs

* Updated documentation

* Fix for hipDataType include

* Addressed review comments

* Minor fixes

+++ clients/include/testing_orgbr_ungbr.hpp
@@ -116,7 +116,7 @@ void orgbr_ungbr_initData(const hipsolverHandle_t   handle,
         using S = decltype(std::real(T{}));
 
         int            info;
-        size_t         s = max(hIpiv.n(), 2);
+        size_t         s = max(hIpiv.n(), int64_t(2));
         std::vector<S> E(s - 1);
         std::vector<S> D(s);
         std::vector<T> P(s);
